Fix rounding time at MethodNode->MethodInfo conversion

+++ Tracer/Tracer.Core/MethodNode.cs
@@ -31,6 +31,6 @@ internal class MethodNode
             methods.Add(child.ToMethodInfo());
         }
 
-        return new(Name, Class, Watch.Elapsed.TotalMilliseconds + "ms", methods);
+        return new(Name, Class, (int)Watch.Elapsed.TotalMilliseconds + "ms", methods);
     }
 }
